requirements.txt
# requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-generativeai
requests
redis